refactor: extract to components
