+++ components/ShiftCard.tsx
@@ -1,0 +1,39 @@
+import styles from "@/styles/ShiftCard.module.css";
+import { identifyShift, toRelativeTime, toTime } from "@/lib/utils";
+import { Card, Divider, Text } from "@geist-ui/core";
+import { PageObjectResponse } from "@notionhq/client/build/src/api-endpoints";
+import { Participents } from "./Participants";
+
+interface IProps {
+  shift: PageObjectResponse;
+  userId: string;
+  allUsers: Array<any>;
+}
+
+export const ShiftCard = ({ shift, userId, allUsers }: IProps) => {
+  const time =
+    shift.properties["זמן"].type == "date" && shift.properties["זמן"].date;
+  const type = identifyShift(shift, userId);
+  const timeString =
+    time &&
+    `${toTime(time.end as string)} - ${toTime(time.start)} (${toRelativeTime(
+      time.start
+    )})`;
+
+  return (
+    <Card width="80%" type="success">
+      <Card.Content className={`${styles.cardHeader}`}>
+        <Text b my={0}>
+          המשמרת הבאה
+        </Text>
+        <Text>{type.emoji}</Text>
+      </Card.Content>
+      <Divider h="1px" my={0} />
+      <Card.Content>
+        <Text h3>{type.name}</Text>
+        <Text h5>{timeString}</Text>
+        <Participents shift={shift} allUsers={allUsers} />
+      </Card.Content>
+    </Card>
+  );
+};
